ganomaly/requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
idna==2.8
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
olefile==0.46
Pillow>=7.1.0
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.2
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
torchfile
tornado==6.0.3
tqdm==4.33.0
urllib3==1.25.3
visdom==0.1.8.8
websocket-client==0.56.0
